pyproject.toml
[tool.poetry]
name = "sheet2json"
version = "0.1.0"
description = "Script to convert receipt data from spreadsheet format to JSON"
authors = ["Artiom Khandamirov <t9max@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
pyexcel = "^0.5.15"
pyexcel-io = "^0.5.20"
pyexcel-xlsx = "^0.5.7"
dnspython = "^1.16"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
flake8-docstrings = "^1.5.0"
flake8-bandit = "^2.1.2"
flake8-deprecated = "^1.3"
flake8-awesome = "^1.3.0"
autopep8 = "^1.5.3"
rope = "^0.17.0"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
